+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Next Release
 
-- Adds new billing functionality for ReferralCustomer users
+- Adds new beta billing functionality for ReferralCustomer users
   - `add_payment_method` can add a pre-existing Stripe bank account or credit card to your EasyPost account
   - `refund_by_amount` refunds your wallet by a dollar amount
   - `refund_by_payment_log` refunds you wallet by a PaymentLog ID